refactored code reviewed by team
