+++ graphql/models/vote.js
