--- graphql/schema/index.js
@@ -1,149 +1,0 @@
-const { buildSchema, GraphQLObjectType, GraphQLString, GraphQLList, GraphQLSchema } = require('graphql');
-
-module.exports = buildSchema(`
-type Event {
-  _id: ID!
-  eventName: String!
-  eventDate: String!
-  eventClosingDate: String!
-  eventLat: String!
-  eventLong: String!
-  eventDistance: String!
-  eventOrganiser: String!
-  attendees: [String!]!
-  restaurantList: ID!
-  restaurants: [RestaurantTA]
-  votesRefs : [ID] 
-  votes: [ID]   
-  winner: String
-  topThree: String
-} 
-
-type Vote{
-  _id: ID
-  eventRef: ID!
-  restaurantRef: ID!
-  positiveVote: Int!
-  negativeVote: Int!
-}
-
-type VoteBox {
-  _id: ID!
-  votes: [Vote!]!
-}
-
-type User {
-  _id: ID!
-  email: String!
-  password: String!
-  firstName: String!
-  city: String!
-}
-
-type RestaurantTA{
-  _id: ID
-  location_id: String
-  location_string: String
-  name: String
-  description: String
-  cuisine: [String]
-  photo: String
-  price: String
-  ranking: String
-  rating: String
-  phone: String
-  website: String
-  address: String
-  dietary_restrictions: [String]
-  num_reviews: String
-  positiveVotes: Int!
-  negativeVotes: Int!
-}
-
-type RestaurantList{
-  _id: ID
-  list: [ID]
-}
-
-input RestaurantTAInput{
-  location_id: String
-  location_string: String
-  name: String
-  description: String
-  cuisine: [String]
-  photo: String
-  price: String
-  ranking: String
-  rating: String
-  phone: String
-  website: String
-  address: String
-  dietary_restrictions: [String]
-}
-
-input EventInput{
-  eventName: String!
-  eventDate: String!
-  eventClosingDate: String!
-  eventLat: String!
-  eventLong: String!
-  eventDistance: String
-  eventOrganiser: String!
-  attendees: [String!]
-}
-
-input UserInput{
-  email: String!
-  password: String!
-  firstName: String!
-  city: String!
-}
-
-input RestaurantListInput{
-  list: [ID]
-}
-
-input TripAdvisorInput {
-  distance: String
-  latitude: String!
-  longitude: String!
-}
-
-input VoteInput {
-  eventRef: ID!
-  restaurantRef: ID!
-  positiveVote: Int!
-  negativeVote: Int!
-}
-
-input VoteBoxInput {
-  votes: String
-}
-
-type RootQuery{
-  getEvents: [Event!]!
-  getEventByID(eventID: ID!): Event!
-  getUsers: [User!]!
-  getVotesByEventID(eventID: ID!): [Vote!]!
-  getRestaurantsTripAdvisor(tripAdvisorInput: TripAdvisorInput ): [RestaurantTA]
-  getRestaurantTA(restaurantID: ID!): RestaurantTA
-  getRestaurantList(listID: ID!): RestaurantList
-  getWinner(eventID: ID!): RestaurantTA
-}
-
-type RootMutation{
-  createEvent(eventInput: EventInput): Event
-  createUser(userInput: UserInput): User
-  createRestaurantTA(restaurantTAInput: RestaurantTAInput): RestaurantTA
-  createRestaurantList(restaurantListInput: RestaurantListInput): RestaurantList
-  createVote(voteInput: VoteInput): Vote
-  createVoteBox(voteBoxInput: VoteBoxInput): Event
-}
-
-schema{
-  query: RootQuery
-  mutation: RootMutation
-}
-
-`)
-
