--- models/user.js
@@ -1,20 +1,0 @@
-const mongoose = require('mongoose');
-
-const Schema = mongoose.Schema;
-
-const userSchema = new Schema({
-  email: {
-    type: String,
-    required: true
-  },
-  firstName: {
-    type: String,
-    required: true
-  },
-  city: {
-    type: String,
-    required: true,
-  }
-});
-
-module.exports = mongoose.model('User', userSchema)
